test: replace unit test with e2e test for issue#375 in openapi.test.ts
